LayoutTests/fast/events/spatial-navigation/snav-iframe-no-focusable-content-expected.txt

a


e
PASS gFocusedDocument.activeElement.getAttribute("id") is "end"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
